Sublime setup fixed.

+++ .misc/setup.py
@@ -6,9 +6,9 @@ def print_intro(msg):
 def print_done():
     print('Done.\n')
 
-SUBLIME_SETTINGS_PATH = ''
+paths = {'SUBLIME_SETTINGS': ''}
 if sys.platform in ('win32', 'cygwin'):
-    SUBLIME_SETTINGS_PATH = os.getenv('HOMEPATH') + '/AppData/Roaming/Sublime Text 3'
+    paths['SUBLIME_SETTINGS'] = 'c:' + os.getenv('HOMEPATH') + '/AppData/Roaming/Sublime Text 3'
 ZSH_INSTALLED = True if shutil.which('zsh') is not None else False
 CURL_INSTALLED = True if shutil.which('curl') is not None else False
 WGET_INSTALLED = True if shutil.which('wget') is not None else False
@@ -21,22 +21,26 @@ print_intro('Sublime Text 3')
 subl_fnames = os.listdir('sublime/')
 subl_settings = 'Preferences.sublime-settings'
 confirm = 'n'
-if subl_settings in os.listdir(SUBLIME_SETTINGS_PATH + '/Packages/User'):
-    print('The settings file is already exists.')
-    confirm = input('Override? [y/N]: ')
-    if confirm not in ('y', 'Y'):
-        confirm = 'n' # Default value is `N`.
+if os.path.isdir(paths['SUBLIME_SETTINGS']):
+    if subl_settings in os.listdir(paths['SUBLIME_SETTINGS'] + '/Packages/User'):
+        print('The settings file is already exists.')
+        confirm = input('Override? [y/N]: ')
+        if confirm not in ('y', 'Y'):
+            confirm = 'n' # Default value is `N`.
+    else:
+        confirm = input('Install? [Y/n]: ')
+        if confirm not in ('n', 'N'):
+            confirm = 'y' # Default value is `Y`.
+    if confirm.lower() == 'y':
+        print('Copying settings file ...')
+        for subl_fname in subl_fnames:
+            shutil.copy('./sublime/' + subl_fname,
+                paths['SUBLIME_SETTINGS'] + '/Packages/User/' + subl_fname)
+        print_done()
+    else:
+        print_done()
 else:
-    confirm = input('Install? [Y/n]: ')
-    if confirm not in ('n', 'N'):
-        confirm = 'y' # Default value is `Y`.
-if confirm.lower() == 'y':
-    print('Copying settings file ...')
-    for subl_fname in subl_fnames:
-        shutil.copy('./sublime/' + subl_fname,
-            SUBLIME_SETTINGS_PATH + '/Packages/User/' + subl_fname)
-    print_done()
-else:
+    print('It seems Sublime Text 3 is not installed in this system.')
     print_done()
 
 # Oh My Zsh.
